applied changes based on PR reviews.
